Football_compition_dataset in json
This dataset provides a comprehensive overview of various sports competitions from around the world. It includes data on football (soccer) tournaments, leagues, and cup matches, covering details such as competition names, locations, types, current seasons, and historical records. With information on multiple seasons and updates, this dataset serves as a valuable resource for sports enthusiasts, analysts, and researchers interested in the global landscape of sports competitions